add `themeBuilder` to theme addon
