--- packages/widgetbook/lib/src/addons/cupertino_theme_addon/addon.dart
@@ -1,2 +1,0 @@
-export './cupertino_theme_addon.dart';
-export './cupertino_theme_setting.dart';
