--- packages/widgetbook/lib/src/addons/cupertino_theme_addon/cupertino_theme_setting.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/cupertino.dart';
-import 'package:widgetbook/widgetbook.dart';
-
-class CupertinoThemeSetting extends ThemeSetting<CupertinoThemeData> {
-  CupertinoThemeSetting({
-    required super.themes,
-    required super.activeTheme,
-  });
-}
