--- packages/widgetbook/lib/src/addons/custom_theme_addon/addon.dart
@@ -1,2 +1,0 @@
-export './custom_theme_addon.dart';
-export './custom_theme_setting.dart';
