--- packages/widgetbook/lib/src/addons/custom_theme_addon/custom_theme_setting.dart
@@ -1,8 +1,0 @@
-import 'package:widgetbook/widgetbook.dart';
-
-class CustomThemeSetting<T> extends ThemeSetting<T> {
-  CustomThemeSetting({
-    required super.themes,
-    required super.activeTheme,
-  });
-}
