--- packages/widgetbook/lib/src/addons/material_theme_addon/addon.dart
@@ -1,2 +1,0 @@
-export './material_theme_addon.dart';
-export './material_theme_setting.dart';
