--- packages/widgetbook/lib/src/addons/material_theme_addon/material_theme_setting.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:widgetbook/widgetbook.dart';
-
-class MaterialThemeSetting extends ThemeSetting<ThemeData> {
-  MaterialThemeSetting({
-    required super.themes,
-    required super.activeTheme,
-  });
-}
